motoblog/requirements.txt
Asgiref==3.8.1
Beautifulsoup4==4.13.4
Django==5.2
Django-taggit==6.1.0
Flake8==7.2.0
Isort==6.0.1
Markdown==3.8
Psycopg2==2.9.10
Python-dotenv==1.1.0
Sqlparse==0.5.3
Tzdata==2025.2
